Fix for new installments.
Fixed so that new installments will use the new naming right away.
But upgrades will migrate the table names.

+++ migrations/2020_14_08_165232_update_table_names.php
@@ -13,25 +13,27 @@ class UpdateTableNames extends Migration
      */
     public function up()
     {
-        Schema::table('transaction_logs', function (Blueprint $table) {
-            $table->dropForeign(['transaction_id']);
-        });
+        if (Schema::hasTable('transactions') && Schema::hasTable('transaction_logs') && Schema::hasTable('merchants')) {
+            Schema::table('transaction_logs', function (Blueprint $table) {
+                $table->dropForeign(['transaction_id']);
+            });
 
-        Schema::table('merchantables', function (Blueprint $table) {
-            $table->dropForeign(['merchant_id']);
-        });
+            Schema::table('merchantables', function (Blueprint $table) {
+                $table->dropForeign(['merchant_id']);
+            });
 
-        Schema::rename('transactions', 'omnipay_transactions');
-        Schema::rename('transaction_logs', 'omnipay_transaction_logs');
-        Schema::rename('merchants', 'omnipay_merchants');
+            Schema::rename('transactions', 'omnipay_transactions');
+            Schema::rename('transaction_logs', 'omnipay_transaction_logs');
+            Schema::rename('merchants', 'omnipay_merchants');
 
-        Schema::table('omnipay_transaction_logs', function (Blueprint $table) {
-            $table->foreign('transaction_id')->references('id')->on('omnipay_transactions')->onDelete('cascade');
-        });
+            Schema::table('omnipay_transaction_logs', function (Blueprint $table) {
+                $table->foreign('transaction_id')->references('id')->on('omnipay_transactions')->onDelete('cascade');
+            });
 
-        Schema::table('merchantables', function (Blueprint $table) {
-            $table->foreign('merchant_id')->references('id')->on('omnipay_merchants')->onDelete('cascade');
-        });
+            Schema::table('merchantables', function (Blueprint $table) {
+                $table->foreign('merchant_id')->references('id')->on('omnipay_merchants')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -41,24 +43,6 @@ class UpdateTableNames extends Migration
      */
     public function down()
     {
-        Schema::table('omnipay_transaction_logs', function (Blueprint $table) {
-            $table->dropForeign(['transaction_id']);
-        });
-
-        Schema::table('merchantables', function (Blueprint $table) {
-            $table->dropForeign(['merchant_id']);
-        });
-
-        Schema::rename('omnipay_transactions', 'transactions');
-        Schema::rename('omnipay_transaction_logs', 'transaction_logs');
-        Schema::rename('omnipay_merchants', 'merchants');
-
-        Schema::table('transaction_logs', function (Blueprint $table) {
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
-        });
-
-        Schema::table('merchantables', function (Blueprint $table) {
-            $table->foreign('merchant_id')->references('id')->on('merchants')->onDelete('cascade');
-        });
+        // There is no coming back anymore from here.
     }
 }
